specify dtype

+++ vae.py
@@ -15,7 +15,8 @@ class VideoAutoEncoderKL(nn.Module):
                  **kwargs) -> None:
         super().__init__(*args, **kwargs)
 
-        self.vae = AutoencoderKL.from_pretrained(pretrained_model)
+        self.vae = AutoencoderKL.from_pretrained(pretrained_model,
+                                                 torch_dtype=torch.float)
         self.micro_batch_size = micro_batch_size
         self.patch_size = patch_size
